+++ examples/data-sources/nxos_ipv4_access_list_policy_ingress_interface_instance/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_ipv4_access_list_policy_ingress_interface_instance" "example" {
+  interface_id = "eth1/10"
+}