+++ examples/resources/nxos_ipv4_access_list_policy_ingress_interface_instace/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_ipv4_access_list_policy_ingress_interface_instace.example "sys/acl/ipv4/policy/ingress/intf-[eth1/10]/acl"