+++ internal/provider/data_source_nxos_ipv4_access_list_policy_ingress_interface_instance.go
@@ -16,23 +16,23 @@ import (
 
 // Ensure the implementation satisfies the expected interfaces.
 var (
-	_ datasource.DataSource              = &IPv4AccessListPolicyIngressInterfaceInstaceDataSource{}
-	_ datasource.DataSourceWithConfigure = &IPv4AccessListPolicyIngressInterfaceInstaceDataSource{}
+	_ datasource.DataSource              = &IPv4AccessListPolicyIngressInterfaceInstanceDataSource{}
+	_ datasource.DataSourceWithConfigure = &IPv4AccessListPolicyIngressInterfaceInstanceDataSource{}
 )
 
-func NewIPv4AccessListPolicyIngressInterfaceInstaceDataSource() datasource.DataSource {
-	return &IPv4AccessListPolicyIngressInterfaceInstaceDataSource{}
+func NewIPv4AccessListPolicyIngressInterfaceInstanceDataSource() datasource.DataSource {
+	return &IPv4AccessListPolicyIngressInterfaceInstanceDataSource{}
 }
 
-type IPv4AccessListPolicyIngressInterfaceInstaceDataSource struct {
+type IPv4AccessListPolicyIngressInterfaceInstanceDataSource struct {
 	data *NxosProviderData
 }
 
-func (d *IPv4AccessListPolicyIngressInterfaceInstaceDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_ipv4_access_list_policy_ingress_interface_instace"
+func (d *IPv4AccessListPolicyIngressInterfaceInstanceDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_ipv4_access_list_policy_ingress_interface_instance"
 }
 
-func (d *IPv4AccessListPolicyIngressInterfaceInstaceDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
+func (d *IPv4AccessListPolicyIngressInterfaceInstanceDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
 		MarkdownDescription: helpers.NewResourceDescription("This data source can read an IPv4 Access List Policy Ingress Interface Instance.", "aclInst", "Security%20and%20Policing/acl:Inst/").String,
@@ -58,7 +58,7 @@ func (d *IPv4AccessListPolicyIngressInterfaceInstaceDataSource) Schema(ctx conte
 	}
 }
 
-func (d *IPv4AccessListPolicyIngressInterfaceInstaceDataSource) Configure(_ context.Context, req datasource.ConfigureRequest, _ *datasource.ConfigureResponse) {
+func (d *IPv4AccessListPolicyIngressInterfaceInstanceDataSource) Configure(_ context.Context, req datasource.ConfigureRequest, _ *datasource.ConfigureResponse) {
 	if req.ProviderData == nil {
 		return
 	}
@@ -66,8 +66,8 @@ func (d *IPv4AccessListPolicyIngressInterfaceInstaceDataSource) Configure(_ cont
 	d.data = req.ProviderData.(*NxosProviderData)
 }
 
-func (d *IPv4AccessListPolicyIngressInterfaceInstaceDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
-	var config, state IPv4AccessListPolicyIngressInterfaceInstace
+func (d *IPv4AccessListPolicyIngressInterfaceInstanceDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
+	var config, state IPv4AccessListPolicyIngressInterfaceInstance
 
 	// Read config
 	diags := req.Config.Get(ctx, &config)